Even more tweaking

+++ src/fire-engine.h
@@ -5,12 +5,12 @@
 
 typedef struct
 {
-    int width;
-    int height;
-    int *data;
+    uint16_t width;
+    uint16_t height;
+    uint8_t *data;
 } doom_fire_buffer_t;
 
-void engine_create_buffer(int width, int height, doom_fire_buffer_t **buffer);
+void engine_create_buffer(uint16_t width, uint16_t height, doom_fire_buffer_t **buffer);
 void engine_destroy_buffer(doom_fire_buffer_t **buffer);
 void engine_init_buffer(doom_fire_buffer_t *buffer, uint8_t ignition_value);
 void engine_step_fire(doom_fire_buffer_t *buffer);
